split out world tests for the event system
