+++ ecs/test_types_test.go
@@ -1,0 +1,13 @@
+package ecs_test
+
+// Position component.
+type Position struct {
+	X int
+	Y int
+}
+
+// Velocity component.
+type Velocity struct {
+	X int
+	Y int
+}
